Move transformation into its own module
